requirements.txt
# Python 3.10.12 required
django
djangorestframework
pillow
psycopg2-binary
python-decouple
djangorestframework-simplejwt
django-debug-toolbar
opencv-python
opencv_contrib-python
face_recognition
numpy
celery
redis
django-celery-beat